requirements.txt
beautifulsoup4~=4.9.3
selenium~=3.141.0
mysql-connector-python~=8.0.21